fix statistic out of integer range

+++ includes/SearchEngine.h
@@ -68,7 +68,7 @@ public:
 	PNSEngine* pnsEngine;
 
 	//for statistic
-	int totalSearchNode;
+	long totalSearchNode;
 
 	// change player
 	__forceinline void ChangeSide()
